refactor: MemoDataHolder에서 MemoData로 이름 수정
