+++ CloudNotes/CloudNotes/Data/MemoDataHolder.swift
@@ -1,25 +1,8 @@
 //
-//  DataHolder.swift
+//  MemoDataTransfer.swift
 //  CloudNotes
 //
-//  Created by Do Yi Lee on 2021/09/02.
+//  Created by Do Yi Lee on 2021/09/09.
 //
 
 import UIKit
-
-class MemoDataHolder {
-    static var list = generateList()
-    
-    private init() { }
-    
-    private static func generateList() -> [MemoDecodeModel]? {
-        guard let dataAsset = NSDataAsset(name: "sample")?.data else { return nil }
-        
-        do {
-            let decodedData = try JSONParser.decoder(modelType: [MemoDecodeModel].self, jsonData: dataAsset)
-            return decodedData
-        } catch {
-            return nil
-        }
-    }
-}
